refactor: split identities utils client and server
